requirements.txt
gymnasium[box2d,mujoco]
torch
numpy
wandb
pyyaml
imageio
setuptools